Moving code into versioned api
